Add files via upload
Stores numbers into an array a and sort it - a.sort(). Uses the linked list and node classes o store the ordered elements of a into a LinkedList structure L. Ask the user a Integer value x. Looks for the position to insert x in L. If the value x is already in L, remove it. If it is not, insert x in the appropriate position so L remains sorted.